requirements.txt
setuptools
beautifulsoup4
requests
python-emailahoy3
pyexcel==0.2.1
pyexcel_ods==0.1.1
pyexcel_xls==0.1.0
pyexcel_xlsx==0.1.0
pyexcel_io==0.1.0
pyexcel_text==0.2.0
tweepy
networkx
decorator
validate_email
tabulate
oauthlib>=1.0.0
python-whois
pyyaml
colorama
configparser
cloudscraper
duckpy
packaging
